+++ frontend/src/lib/components/propertyPicker/ObjectViewer.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { copyToClipboard, pluralize, truncate } from '$lib/utils'
+	import { copyToClipboard, emptyString, pluralize, truncate } from '$lib/utils'
 
 	import { createEventDispatcher } from 'svelte'
 	import { Badge } from '../common'
@@ -7,8 +7,10 @@
 	import WarningMessage from './WarningMessage.svelte'
 	import { NEVER_TESTED_THIS_FAR } from '../flows/models'
 	import Portal from 'svelte-portal'
-	import { PanelRightOpen } from 'lucide-svelte'
+	import { Download, PanelRightOpen } from 'lucide-svelte'
 	import S3FilePicker from '../S3FilePicker.svelte'
+	import { HelpersService } from '$lib/gen'
+	import { workspaceStore } from '$lib/stores'
 
 	export let json: any
 	export let level = 0
@@ -53,6 +55,18 @@
 			copyToClipboard(valueToCopy)
 		}
 		dispatch('select', rawKey ? key : computeKey(key, isArray, currentPath))
+	}
+
+	async function downloadS3File(fileKey: string | undefined) {
+		if (emptyString(fileKey)) {
+			return
+		}
+		const downloadUrl = await HelpersService.generateDownloadUrl({
+			workspace: $workspaceStore!,
+			fileKey: fileKey!
+		})
+		console.log('download URL ', downloadUrl.download_url)
+		window.open(downloadUrl.download_url, '_blank')
 	}
 
 	$: keyLimit = isArray ? 1 : 100
@@ -135,6 +149,13 @@
 			{#if level == 0 && topBrackets}
 				<span class="h-0">{closeBracket}</span>
 				{#if getTypeAsString(json) === 's3object'}
+					<button
+						class="text-secondary underline text-2xs whitespace-nowrap"
+						on:click={() => {
+							downloadS3File(json?.s3)
+						}}
+						><span class="flex items-center gap-1"><Download size={12} />download</span>
+					</button>
 					<button
 						class="text-secondary underline text-2xs whitespace-nowrap ml-1"
 						on:click={() => {